requirements.txt
flask==3.0.0              # Web framework
flask-cors==4.0.0         # Cross-Origin Resource Sharing
google-genai==1.0.0       # Google Gemini AI SDK
requests==2.31.0          # HTTP library
beautifulsoup4==4.12.0    # Web scraping
python-dotenv==1.0.0      # Environment variable management